remove description col
