--- src/components/Create/components/pages/ReviewDeploy/components/DescriptionCol/index.ts
@@ -1,1 +1,0 @@
-export * from './DescriptionCol'
