+++ src/components/Create/components/pages/ReviewDeploy/components/FundingConfigurationReview/FundingConfigurationReview.tsx
@@ -2,7 +2,7 @@ import { t, Trans } from '@lingui/macro'
 import { Tooltip } from 'antd'
 import { twMerge } from 'tailwind-merge'
 import { PayoutsList } from '../../../PayoutsPage/components/PayoutsList'
-import { ReviewDescription } from '../DescriptionCol'
+import { ReviewDescription } from '../ReviewDescription'
 import { useFundingConfigurationReview } from './hooks/FundingConfigurationReview'
 
 export const FundingConfigurationReview = () => {
